+++ neoforge/src/main/java/io/github/wouink/furnish/neoforge/FurnishForge.java
@@ -1,8 +1,12 @@
 package io.github.wouink.furnish.neoforge;
 
 import io.github.wouink.furnish.Furnish;
+import io.github.wouink.furnish.client.gui.ConditionalSlotContainerScreen;
+import io.github.wouink.furnish.client.gui.DiskRackScreen;
 import io.github.wouink.furnish.client.gui.FurnitureWorkbenchScreen;
-import io.github.wouink.furnish.client.renderer.SeatRenderer;
+import io.github.wouink.furnish.client.renderer.*;
+import io.github.wouink.furnish.event.AddArmsToArmorStand;
+import io.github.wouink.furnish.event.CyclePainting;
 import io.github.wouink.furnish.setup.FurnishRegistries;
 import net.neoforged.bus.api.IEventBus;
 import net.neoforged.bus.api.SubscribeEvent;
@@ -12,6 +16,8 @@ import net.neoforged.fml.event.lifecycle.FMLClientSetupEvent;
 import net.neoforged.fml.event.lifecycle.FMLCommonSetupEvent;
 import net.neoforged.neoforge.client.event.EntityRenderersEvent;
 import net.neoforged.neoforge.client.event.RegisterMenuScreensEvent;
+import net.neoforged.neoforge.common.NeoForge;
+import net.neoforged.neoforge.event.entity.player.PlayerInteractEvent;
 
 @Mod(Furnish.MODID)
 @EventBusSubscriber(bus = EventBusSubscriber.Bus.MOD)
@@ -25,8 +31,8 @@ public class FurnishForge {
 
         Furnish.init();
 
-        // Architectury InteractionEvent.INTERACT_ENTITY does not seem to work with Armor Stands and Paintings...
-        //NeoForge.EVENT_BUS.register(FurnishForgeEvents.class);
+        NeoForge.EVENT_BUS.addListener(FurnishForge::rightClickPainting);
+        NeoForge.EVENT_BUS.addListener(FurnishForge::rightClickArmorStand);
     }
 
     private void init(final FMLCommonSetupEvent event) {
@@ -39,15 +45,36 @@ public class FurnishForge {
     }
 
     @SubscribeEvent
-    private static void registerScreens(RegisterMenuScreensEvent event) {
+    public static void registerScreens(RegisterMenuScreensEvent event) {
         // Registering through Architectury does not work anymore in 1.21
-        Furnish.LOG.info("RegisterMenuScreensEvent");
         event.register(FurnishRegistries.Furniture_Workbench_Container.get(), FurnitureWorkbenchScreen::new);
+        event.register(FurnishRegistries.Crate_Container.get(), ConditionalSlotContainerScreen::new);
+        event.register(FurnishRegistries.Mailbox_Container.get(), ConditionalSlotContainerScreen::new);
+        event.register(FurnishRegistries.Bookshelf_Chest_Container.get(), ConditionalSlotContainerScreen::new);
+        event.register(FurnishRegistries.Disk_Rack_Container.get(), DiskRackScreen::new);
+        Furnish.LOG.info("Bound Furnish containers to their screens (special Forge 1.21 workaround).");
     }
 
     @SubscribeEvent
-    private static void registerEntityRenderer(EntityRenderersEvent.RegisterRenderers event) {
+    public static void registerEntityRenderer(EntityRenderersEvent.RegisterRenderers event) {
         // Registering through Architectury does not work anymore in 1.21
         event.registerEntityRenderer(FurnishRegistries.Seat_Entity.get(), SeatRenderer::new);
+        event.registerBlockEntityRenderer(FurnishRegistries.Mailbox_BlockEntity.get(), MailboxRenderer::new);
+        event.registerBlockEntityRenderer(FurnishRegistries.Plate_BlockEntity.get(), PlateRenderer::new);
+        event.registerBlockEntityRenderer(FurnishRegistries.Shelf_BlockEntity.get(), ShelfRenderer::new);
+        event.registerBlockEntityRenderer(FurnishRegistries.Showcase_BlockEntity.get(), ShowcaseRenderer::new);
+        event.registerBlockEntityRenderer(FurnishRegistries.Disk_Rack_BlockEntity.get(), DiskRackRenderer::new);
+        event.registerBlockEntityRenderer(FurnishRegistries.Recycle_Bin_BlockEntity.get(), RecycleBinRenderer::new);
+        Furnish.LOG.info("Registered Furnish entity/block entity renderers (special Forge 1.21 workaround).");
+    }
+
+    public static void rightClickArmorStand(PlayerInteractEvent.EntityInteractSpecific event) {
+        // Architectury InteractionEvent.INTERACT_ENTITY does not seem to work with Armor Stands and Paintings...
+        AddArmsToArmorStand.rightClickArmorStand(event.getEntity(), event.getTarget(), event.getHand());
+    }
+
+    public static void rightClickPainting(PlayerInteractEvent.EntityInteract event) {
+        // Architectury InteractionEvent.INTERACT_ENTITY does not seem to work with Armor Stands and Paintings...
+        CyclePainting.onPaintingInteract(event.getEntity(), event.getTarget(), event.getHand());
     }
 }